brcmfmac: fix compile error on v4.x kernel and 32bit system
- v5.15.58-backports/drivers/net/wireless/broadcom/brcm80211/brcmfmac/bt_shared_sdio_ifx.c:276:12: error: format ‘%lld’ expects argument of type ‘long long int’, but argument 4 has type ‘__kernel_time_t {aka long int}’ [-Werror=format=]

- v5.15.58-backports/drivers/net/wireless/broadcom/brcm80211/brcmfmac/cfg80211.c:832:2: error: ignoring return value of ‘strscpy’, declared with attribute warn_unused_result [-Werror=unused-result]
  strscpy(ie_setbuf->cmd, "add", sizeof(ie_setbuf->cmd));
  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

- v5.15.58-backports/drivers/net/wireless/broadcom/brcm80211/brcmfmac/cfg80211.c:6682:13: error: cast from pointer to integer of different size [-Werror=pointer-to-int-cast]
   *cookie = (u64)mf_params->data;

Fixes SWWLAN-153447
Fixes SWWLAN-154232

+++ drivers/net/wireless/broadcom/brcm80211/brcmfmac/bt_shared_sdio_ifx.c
@@ -239,7 +239,7 @@ static void inf_btsdio_err_reg_record(struct inf_bt_if *bt_if,
 	reg->addr = addr;
 	reg->val = val;
 	brcmf_err("[%5lld.%06ld] %8s err: %d\taddr: 0x%x\tval: 0x%x\n",
-		  cmd->time.tv_sec, cmd->time.tv_nsec / NSEC_PER_USEC,
+		  (long long)cmd->time.tv_sec, cmd->time.tv_nsec / NSEC_PER_USEC,
 		  inf_btsdio_err_char(cmd->type), err, addr, val);
 
 	inf_btsdio_err_enq(bt_if, cmd);
@@ -262,7 +262,7 @@ static void inf_btsdio_err_buf_record(struct inf_bt_if *bt_if,
 	buf = &cmd->u.buf;
 	buf->nbytes = nbytes;
 	brcmf_err("[%5lld.%06ld] %8s err: %d\tnbytes: %d\n",
-		  cmd->time.tv_sec, cmd->time.tv_nsec / NSEC_PER_USEC,
+		  (long long)cmd->time.tv_sec, cmd->time.tv_nsec / NSEC_PER_USEC,
 		  inf_btsdio_err_char(cmd->type), err, nbytes);
 
 	inf_btsdio_err_enq(bt_if, cmd);
@@ -287,7 +287,7 @@ static void inf_btsdio_err_mem_record(struct inf_bt_if *bt_if,
 	mem->addr = addr;
 	mem->size = size;
 	brcmf_err("[%5lld.%06ld] %8s err: %d\tset: %d\taddr: 0x%x\tsize: %d\n",
-		  cmd->time.tv_sec, cmd->time.tv_nsec / NSEC_PER_USEC,
+		  (long long)cmd->time.tv_sec, cmd->time.tv_nsec / NSEC_PER_USEC,
 		  inf_btsdio_err_char(cmd->type), err, set, addr, size);
 
 	inf_btsdio_err_enq(bt_if, cmd);
@@ -312,7 +312,7 @@ static void inf_btsdio_err_dump(struct seq_file *seq, struct inf_bt_if *bt_if)
 	spin_lock(&bt_if->err_list_lock);
 	list_for_each_entry(cmd, &bt_if->err_list, list) {
 		seq_printf(seq, "%3d: [%5lld.%06ld] %8s err: %d\t",
-			   ++idx, cmd->time.tv_sec, cmd->time.tv_nsec / NSEC_PER_USEC,
+			   ++idx, (long long)cmd->time.tv_sec, cmd->time.tv_nsec / NSEC_PER_USEC,
 			   inf_btsdio_err_char(cmd->type), cmd->err);
 		switch (cmd->type) {
 		case ERR_REG_RB:
